split up analysis.rs
