+++ src-tauri/src/analysis/mod.rs
@@ -1,0 +1,2 @@
+pub mod analysers;
+pub mod commands;
